Fix rng seed for kos15 unit tests
Encountered a failed kos15 unit test, because of some very special rng seed.

The test which failed was `test_ot_splitting_mix_pairs`, which is
expected to panic because it uses a wrong pair of split senders and
receivers. I believe that it failed (did not panic, like expected)
because  the rng seed incidentally produced some blocks, where the first
half does not differ from the second half. This probably happened because
the number of blocks produced is a random number between (1..1024) and I
expect that the number chosen by the rng seed was very small, so that
there was a high probability that this could happen.

We agreed to seed unit tests with a fixed seed, so I took this
opportunity to just do that for the kos15 unit tests.