fix(notary-server): parse JWT claims from env vars
Parsing is facilitated through a custom serde deserializer which
accept both string and struct representation and correctly unpacks
it into JwtClaim struct.

The implementation of the deserializer is generic over the output
type `T`. The only requirement is that `T` implements `std::str::FromStr`
trait and that `std::str::FromStr::Err = eyre::Error`.